Selesai Memindahkan BabyChart File
